+++ src/CLI/usbguard.cpp
@@ -40,6 +40,8 @@
 #include "usbguard-remove-rule.hpp"
 #include "usbguard-watch.hpp"
 #include "usbguard-read-descriptor.hpp"
+#include "usbguard-add-user.hpp"
+#include "usbguard-remove-user.hpp"
 
 namespace usbguard
 {
@@ -57,7 +59,9 @@ namespace usbguard
     { "remove-rule", &usbguard_remove_rule },
     { "generate-policy", &usbguard_generate_policy },
     { "watch", &usbguard_watch },
-    { "read-descriptor", &usbguard_read_descriptor }
+    { "read-descriptor", &usbguard_read_descriptor },
+    { "add-user", &usbguard_add_user },
+    { "remove-user", &usbguard_remove_user }
   };
 
   static void showTopLevelHelp(std::ostream& stream = std::cout)
@@ -82,6 +86,9 @@ namespace usbguard
     stream << "  generate-policy                Generate a rule set (policy) based on the connected USB devices." << std::endl;
     stream << "  watch                          Watch for IPC interface events and print them to stdout." << std::endl;
     stream << "  read-descriptor                Read a USB descriptor from a file and print it in human-readable form." << std::endl;
+    stream << std::endl;
+    stream << "  add-user <name>                Add USBGuard IPC user/group (requires root privilges)" << std::endl;
+    stream << "  remove-user <name>             Remove USBGuard IPC user/group (requires root privileges)" << std::endl;
     stream << std::endl;
   }