+++ src/Tests/Packaging/spell-check.sh
@@ -44,12 +44,14 @@ export LANG="en_US.UTF-8"
 while read path; do
   WORDS="$($PANDOC -t html $path | $ASPELL -l en --add-filter=html $ASPELL_OPTIONS list | sort | uniq -c)"
   if [[ -n "$WORDS" ]]; then
-    echo "ERROR: Found missspelled words in $(realpath --relative-base="$srcdir" "$path"):"
+    echo
+    echo "ERROR: Found missspelled words in $(readlink -f "$path"):"
     echo "========================================"
     echo -e "$WORDS"
     echo "========================================"
     echo "To whitelist a word, add it to this dictionary:"
-    echo "  $(realpath --relative-base="$srcdir" "$EXTRA_DICTIONARY")"
+    echo "  $(readlink -f "$EXTRA_DICTIONARY")"
+    echo
     retval=1
   fi
 done <<EOF